Got the testcases of JmeSystem working. The JmeSystemTest now also tests initialization of the delegates.
Created an AbstractSystemStateDelegate, as it turned out that we still needed one.
Also implemented a missing method (getStorageFolder) on the AndroidSystemIODelegate.
That one needed an activity, so JmeAndroidSystem.getActivity is still in use as dirty quickfix.

Changed it requires changing more code, which means more unit tests.
We don't have the time for that.

Also updated the report to reflect the fact it is all tested now.